@@ -1,14 +1,18 @@
-package com.udacity.pricing.domain.price;
+package com.udacity.pricing.entity;
 
 import java.math.BigDecimal;
+import javax.persistence.*;
 
 /**
  * Represents the price of a given vehicle, including currency.
  */
+@Entity
 public class Price {
 
     private String currency;
     private BigDecimal price;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
     private Long vehicleId;
 
     public Price() {
